Allow to train and validate on different (sets of) datasets.
Modified the episode creation pipeline such that if a dataset has 0 classes for a particular split, that dataset will be ignore when drawing episodes for that split.

Additionally, one can specify a list of datasets to use specifically for validation purposes. This is a new argument to Trainer: validation_dataset_list. If it's empty then the behavior will be the previous one: all datasets in dataset_list are used for training. If ImageNet is part of that list, then only ImageNet is used for validaiton. Otherwise validation happens on (the valid split) of all datasets that are selected for training.

I considered modifying the BenchmarkSpecification directly to add another list that stores for every dataset the list of splits it should contribute to. But the current solution seemed simpler, especially since Trainer.get_benchmark_specification() already returned a different specification for validation purposes.

PiperOrigin-RevId: 256435389