Baseline++ and weight normalization
For each of the kNN- and Finetune- Baselines, add the option to use the cosine classifier of "A Closer Look at Few-shot Classification" during meta-training, and meta-testing for the Finetune baseline (not applicable for meta-testing of the kNN variant).

Also coded weight normalization in the linear classifier layer, since it seems necessary for obtaining good results for Baseline++, and added a hyper to determine whether weight normalization will be enabled. This allows to use it in the context of the non-cosine classifier baselines too.

PiperOrigin-RevId: 263556919