Set the output dimensionality of batch baselines equal to the number of classes in the appropriate split.
The dimensionality of the output layer for the batch baselines was equal to the total number of classes in all datasets, and not just the number of classes pooled across the appropriate split of all datasets. This is vestigial code from a time when class labels were not contiguous (in order to avoid a class ID that exceeds the dimensionality of the output layer). This means that some parameters in the output layer (corresponding to test IDs of meta-validation and meta-testing classes) were never active in meta-training.

PiperOrigin-RevId: 311237044